Temperature changing due to keyLocationLiveData

+++ app/src/main/java/com/pcg/accuweatherapp/model/currentweather/MetricModel.kt
@@ -1,5 +1,5 @@
 package com.pcg.accuweatherapp.model.currentweather
 
 data class MetricModel(
-    val Value: Int
+    val Value: Float
 )
